Add files via upload
Sir, I would like to mention that I have the model.pkl file created in my local. Since GitHub is not allowing files greater than 25MB. I am not able to upload it here. But I have shown it in my demo video and in my Project file. So, Kindly consider it.